agri_chatbot/requirements.txt
flask==2.3.4
openai==1.3.0
joblib==1.3.2
scikit-learn==1.3.2
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.0